+++ src/data/panels.tsx
@@ -1,3 +1,20 @@
 export enum PANELS {
     RECIPE,
 }
+
+export interface IPanelTheme {
+    panelColor: string,
+    barColor: string,
+    buttonHoverColor: string
+}
+
+export const PANEL_THEMES = new Map<PANELS,IPanelTheme>([
+    [
+        PANELS.RECIPE,
+        {
+            panelColor: "var(--color-yellow)",
+            barColor: "var(--color-primary)",
+            buttonHoverColor: "var(--color-secondary)"
+        }
+    ]
+])